FDS Source: radi: Make ZZ_GET an allocatable
This is akin to the prior change in divg. The combination of OpenMP
and array sizes being set at runtime from global variables seems to
mess up the compiler.

Turning it into an allocatable solves this.
I am continouing to research the root cause of this.
ZZ_GET is allocated at the beginning of module so that the behaviour
should be the same as before.


Former-commit-id: 177235cb06965ae6683cb5dfe0081f85a2599f3b [formerly 3a4420d9d8fc752065409d2da24b5bb11ed07533]
Former-commit-id: 879613dfb422c69425cd3c625e6c7d703d699d27
Former-commit-id: 684fbb710051ed7a153eb040ffcecf58c46b7254