FDS Source: radi: Make ZZ_GET an allocatable
This is akin to the prior change in divg. The combination of OpenMP
and array sizes being set at runtime from global variables seems to
mess up the compiler.

Turning it into an allocatable solves this.
I am continouing to research the root cause of this.
ZZ_GET is allocated at the beginning of module so that the behaviour
should be the same as before.