FDS_Source: radi: Parallel radiation spread
Introduces a second version of the radiation propagation.
This version removes the loop-carried dependencies of the original
version and therefore allows parallelisation with OpenMP.

The loop-carried dependencies basically cause a wave to spread
through the mesh, were the cells affected by the wave front are
dependent upon the cells in the slice lying behind them.

To remove the loop-carried dependencies the indices of these slices
are determined and stored in an array.
Subsequently the loop over the cells in each slice can be safely
parellelised with OpenMP.

To ease verification and code review this new version is only used
if USE_OPENMP is .true.


Former-commit-id: 79fa5677aec508bc8b8e7faf7ffd648bfe74ed0b [formerly cf7231a58f9427c6a6ca8d8ae0e49f05c1aa9609]
Former-commit-id: 7ddba2044c652d6d785c8c5e67e85a3d4f32edf9
Former-commit-id: ccd49e38b03fdf4842500c3ad94fa6e793fb0736