FDS Source: radi: Move slice calculation to init.
Determine the wavefront slices for propagation during init_radiation.

The slices over which the cartesian radiation spread is parallelised
were previously calculated during each individual radiation angle.
This created serial overhead.

The SLICE_IJK array stores the wavefront slices for each of the eight
propagation directions.
The propagation direction index is essentially a conversion of the
binary number consisting of the three stepping directions to decimal.
The slices are stored in a serialised fashion since we know the total
number of cells a priori and can allocate an array occordingly.
The size of the individual slices depends on the meshes dimensions, but
remains the same for each direction. Therefore two 1D arrays store the
lower and upper bounds of each slice withing SLICE_IJK.

The propagation can then safely be parallelised within each slice.


Former-commit-id: 5f5a6f69d1f1cb46f626eb2f39ef8e6fdc9167a5 [formerly 54ed6b2fd8edeb671eba2216082e31b1ab8a25d8]
Former-commit-id: 5270675c43f0ee59b828acd4b887ec31008378f6
Former-commit-id: 24bf8adeda59670bbdc0649d7c1d7c08edcbb48a