tkorhon1: evac.f90 revision and compile date removed
The SVN revision and compile date etc are not anymore so easy to get,
so these are taken away from the CHID_evac.out text output. These are
not needed, because you can see them also in CHID.out text file.
They used to be evacid, evacrev, evacdate. So, the svn number of the
evac.f90 was there and its date. (Or something like that, perhaps
these were never working as they should.)