FDS Source: evac.f90 changes and a small evacuation_only mesh in init.f90 change
The SUBROUTINE INIT_WALL_CELL modified so that IIO_MIN etc are also initialized
for evacuation_only(nm) meshes. These initializations were moved a couple of
lines earlier (outside of the if-statement).

Evac.f90 now deals correctly for thin obstructions behind the doors/exits
in the visibility mesh. Now it is check that we are on the correct side
of the door (IOR). Door XB is always on a mesh cell boundary, so it can
be seen on both sides.