smv build: Add mingw/msys as a build target for smokeview.
Mingw and msys are used on Windows systems to utilise gcc for compilation.
This commit adds mingw as a build target. The change is mostly an addition
to the Makefile, and so should have no effect on other builds. Some header
files needed to have stdio.h and options.h included for definitions. These
have been placed under macros to only apply when using mingw.

Where files have been modified, trailing whitespace was removed
automatically. This has been left in as it is consistent with the coding
guidelines.


Former-commit-id: 30ee4874a852b138fc31f2b01451663e7d318214 [formerly 28cbf6cc39597dc7706ac8f0dc12374120948b66]
Former-commit-id: 024971f1d7f79bae6727167f9e640ff073387707
Former-commit-id: 5cb35eb97b64e804a93bb233b4bb90a38a58e0f9