smv build: Add mingw/msys as a build target for smokeview.
Mingw and msys are used on Windows systems to utilise gcc for compilation.
This commit adds mingw as a build target. The change is mostly an addition
to the Makefile, and so should have no effect on other builds. Some header
files needed to have stdio.h and options.h included for definitions. These
have been placed under macros to only apply when using mingw.

Where files have been modified, trailing whitespace was removed
automatically. This has been left in as it is consistent with the coding
guidelines.