Evacuation: Source code fixes, new Evac version 2.5.2 number
Some small bugs corrected in evac.f90. The version is now 2.5.2
due to these bug fixes. No other changes in the algorithms etc.
So, the results of version 2.5.2 should be the same as version
2.5.1, if the bug corrections do not affect the calculation. The
bug fixes were minor fixes for the "see_door" and "see_each_other"
and simular things. Nothing was changed in the "main movement
algorithms".

Also the main.f90 was modified in the evac_main_loop so that
it works now at the last fire time step (Kevin(?) made a change
in the last time step in the main time loop and now also
evacuation time loop does this).