EVACuation related changes to main.f90 (MPI)
Now the combined fire+evacuation MPI calculation
works. EXCHANGE_RADIATION needed some evacuation-IF constructs. These
do not affect plain fire calculation (IF statements check that some
evacuation meshes are present). The problem was, that the fire
processes stop and wait each other at the radiation exchange. And this
exchange is not done at every ICYC. So, the evacuation process should
also stop at these ICYC.

Also an addition (new input keyword) added to evac.f90 so that it is
easier to debug the MPI fire+evac stuff. (Reads/calculates the
CHID_evac.fed file, but does not use the information.)