Update box_burn_away_2D.fds
Correct the mass of the box.  Even though the box is set to 1 m, however the domain is only 0.4 m, where .6 m will be ignored. Therefore, the actual mass that will be compared is within the domain, i.e. 0.4. Hence the total mass is 1.28 kg as found in the Devc file, or by integrating the MLR area.