EVAC device state write to .smv file corrected
Now the evacuation door,entr,exit change is written to the smv file in
the main.f90 routine (same place as device status changes). Earlier
versions these writes were in dump.f90, but at some point devices
write was moved to the main.f90. At this point, the updating the
door,entr,exit status (open or not) for Smokeview was broken.
Now it seems to work.