FDS Source: Evacuation fix for mpi and for the prt5.bnd files
Now also the evacuation meshes has output for the prt5.bnd files.
Now fds is running without error messages, but I do not know if
Smokeview can read the evacuation mesh prt5.bnd files, but this
is an another issue.

Also now the mpi with multiple presesses is fine for fire+evacuation
simulation. The PROCESS(NM) is set =MAX(0,EVAC_PROCESS) for the
evacuation meshes, before it was undefined for evacuation meshes
(or was it zero?).

Both changed do not modife "the fire source code", i.e., they are
done just for the evacuation meshes and/or have "evacuation if"
around the modifications.