FDS+Evac Source: A small addition to evac.f90: OtherTRPE
This was a pull request on tkorhon1/fds. Added to the source, but
not activated there (commented out). Should be tested by Timo to
see that does not brake something that was working. Probably should
add a user input switch so that user can choose: Old way of doing
things or the new way of doing things. Affects "herding agents"
and "lost agents".