FDS Source: Only evacuation related changes to MPI in main.f90
Now things should be working for MPI. The evacuation "phase 2" calculation (only fire meshes, read .xyz fiel, write .fed file) is working for MPI.  The .fed file is written by the "evacuation process" that is defined to be the last process (the last fire mesh proces) for now.  Later it could be changed to process 0. And one can get rid of EVAC_PROCESS, because no fire and evacuation meshes in the same run anymore.