FDS+Evac Source: Small fix to evac mc mode smv file write
Now the evacuation mc mode (using fed) writes the smokeview file
as CHID_evmc.smv. It begins with the fire smv file copied up to
the point where the mesh information ends (last CVENT). The
NMESHES is changed to fire+evacuaion meshes. After this, the
evaucuation stuff is written to the CHID_evmc.smv file. Just
before the end of the run, the rest of the fire smv file (strings)
is appended to the end of the CHID_evmc.file.