FDS+Evac Source: evac.f90 changed for the "no smoke" criterion for doors
The "no smoke" with visibility criterion is now including the distance to the door/exit. The test is now K_ave_agent_to_door*distance_toDoor < K_criterion that is user (default 3/1000) input. The previous version had K_default 3/100, so the new criterion is the same, if the agent is 10 m away from the door. Why this change was done? Without it, an agent was changing the target door even, if it was really close to it.